+++ src/MethodOfLines.jl
@@ -139,7 +139,7 @@ include("precompile.jl")
 
 # Export
 export MOLFiniteDifference, discretize, symbolic_discretize, ODEFunctionExpr, generate_code,
-    edge_align, center_align, get_discrete, chebyspace
+    ode_compile, edge_align, center_align, get_discrete, chebyspace
 export UpwindScheme, WENOScheme, FunctionalScheme, MOLDiscCallback
 
 end